move SimpleImputer to testing module
