--- openml/_backport.py
@@ -1,6 +1,0 @@
-try:
-    from sklearn.impute import SimpleImputer
-except ImportError:
-    from sklearn.preprocessing import Imputer as SimpleImputer
-
-__all__ = ['SimpleImputer']
