+++ tests/test_study/test_study_examples.py
@@ -1,4 +1,4 @@
-from openml.testing import TestBase
+from openml.testing import TestBase, SimpleImputer
 
 
 class TestStudyFunctions(TestBase):
@@ -30,7 +30,7 @@ class TestStudyFunctions(TestBase):
         import sklearn.pipeline
         import sklearn.preprocessing
         import sklearn.tree
-        from openml._backport import SimpleImputer
+
         benchmark_suite = openml.study.get_study(
             'OpenML100', 'tasks'
         )  # obtain the benchmark suite
